Reorganize server code into domain-based packages
