--- src/main/java/com/example/authserver/ProductValidator.java
@@ -1,5 +1,0 @@
-package com.example.authserver;
-
-public interface ProductValidator {
-    boolean isValid(Product product);
-}
