+++ src/main/java/com/example/authserver/http/ProductHttpHandler.java
@@ -1,5 +1,9 @@
-package com.example.authserver;
+package com.example.authserver.http;
 
+import com.example.authserver.http.handler.HandlerFactory;
+import com.example.authserver.http.handler.RouteHandler;
+import com.example.authserver.http.handler.StaticResponseHandler;
+import com.example.authserver.repository.ProductRepository;
 import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpHandler;
