+++ src/main/java/com/example/authserver/http/handler/StaticResponseHandler.java
@@ -1,5 +1,6 @@
-package com.example.authserver;
+package com.example.authserver.http.handler;
 
+import com.example.authserver.http.ResponseWriter;
 import com.sun.net.httpserver.HttpExchange;
 
 import java.io.IOException;
