+++ src/main/java/com/example/authserver/http/handler/product/ListProductsHandler.java
@@ -1,5 +1,7 @@
-package com.example.authserver;
+package com.example.authserver.http.handler.product;
 
+import com.example.authserver.http.ResponseWriter;
+import com.example.authserver.repository.ProductRepository;
 import com.sun.net.httpserver.HttpExchange;
 
 import java.io.IOException;
