+++ src/main/java/com/example/authserver/validation/ProductValidator.java
@@ -1,0 +1,7 @@
+package com.example.authserver.validation;
+
+import com.example.authserver.domain.Product;
+
+public interface ProductValidator {
+    boolean isValid(Product product);
+}
